ai_video_marketing/requirements.txt
# AI Video Marketing System Requirements
# Core AI Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# Image Generation & Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
real-esrgan>=0.3.0

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
opencv-contrib-python>=4.8.0

# Web Scraping & API
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
youtube-dl>=2023.7.6
pytube>=15.0.0

# OpenAI Integration
openai>=0.27.0
tiktoken>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text Processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0

